landing css / logo / refactor
