+++ src/App.css
@@ -1,5 +1,13 @@
-body {
-  margin: 0;
-  background-color: white;
-  height: 100vh;
+html, body {
+    background-color: white;
+    font-family: 'Roboto', sans-serif;
+}
+@media(min-width:1200px) {
+    html, body {
+        margin-left: 15vw;
+        margin-right:15vw;
+        height: 100vh;
+        width: 70vw;
+        max-width: 70vw;
+    }
 }
