+++ src/components/InitialBaner/Style.css
@@ -1,35 +1,37 @@
-div.Header{
-    width: 100vw;
-    max-height: 15vh;
+
+div.Header {
+    width: 100%;
+    max-height: 15%;
     text-align: left;
-    margin-top: 2vh;
+    margin-top: 2%;
+    margin-bottom: 4%;
 }
 div.Header > div{
-    padding-left: 1vh;
-    padding-right: 1vh;
+    padding-left: 1%;
+    padding-right: 1%;
 }
 .Header > .Logo{
     float: left;
 }
 .Header > .Logo > img{
-    max-height: 15vh;
+    max-height: 15%;
 }
 .Header > .Title{
     float: left;
     font-weight: bolder;
     font-size: x-large;
-    padding-top: 2vh;
+    /*padding-top: 2%;*/
 }
 .Header > .Menu{
     float: right;
-    margin-top: 7vh;
+    margin-top: 2em;
 }
 .Header > .Menu ul{
     list-style: none;
 }
 .Header > .Menu ul li{
     display: inline;
-    margin-left: 1vw;
-    margin-right: 1vw;
+    margin-left: 2em;
+    margin-right: 2em;
 }
